User Package was created and User class was moved.
