+++ src/main/java/UserCreator/UserAccountList.java
@@ -1,4 +1,6 @@
 package UserCreator;
+import User.User;
+
 import java.io.Serializable;
 import java.util.HashMap;
 import java.util.Map;
